// src/app/shared/services/api/model/get-all-model-response.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetAllModelResponse { 
    id?: number;
    name?: string;
    brandId?: number;
    brandName?: string;
    fuelId?: number;
    fuelName?: string;
    transmissionId?: number;
    transmissionName?: string;
}